// src/pages/weather/cards/types.d.ts
export interface IResourceByArea {
    KYIV: {
        bonus: {
            full: number
            partial: number
            rest: number
        }
        slots: {
            high: number
            medium: number
            normal: number
            low: number
        }
    }
    SATELLITE: {
        bonus: {
            full: number
            partial: number
            rest: number
        }
        slots: {
            high: number
            medium: number
            normal: number
            low: number
        }
    }
    MIO: {
        bonus: {
            full: number
            partial: number
            rest: number
        }
        slots: {
            high: number
            medium: number
            normal: number
            low: number
        }
    }
    SMALL: {
        bonus: {
            full: number
            partial: number
            rest: number
        }
        slots: {
            high: number
            medium: number
            normal: number
            low: number
        }
    }
}